Fix ability to detect disconnects and reconnect automatically.
Turns out this required two things, one being a fix to the
state-machine, but another being a modification to net_connect to add
the functionality to use the SOL_TCP option TCP_USER_TIMEOUT, which is
what you need when you have something in your send buffer if the
connection breaks. This required an update to the upstream katcp_devel.