Release 1.1.
Add ability to go directly to the Nth array by size across all CMCs via the URL.

This required the addition of the following functions:
- array_get_size
- cmc_server_get_n_arrays

The web_client_handle_requests function is also extended to handle when a client
requests the sorted array-list directly. The cmc_aggregator class is created and
added, which runs after the communication with the CMC servers in the select
loop, grabs the lists of arrays from each cmc server, and sorts them according to
the size of the array, greatest first.