Modified some logic to remove duplicated top-row sensors.
I suspect that this is happening because array_activate() is being
called twice - once when the array becomes "nominal" and again when its
config-file is populated. The logic now should fix that up.

I suspect the same issue is why the first column (input names)
sometimes goes missing - I need to test more, but hopefully this will
fix the problem.